pricebot/requirements.txt
requests
holidays
python-dotenv
py-telegram-notifier
beautifulsoup4